Retry an expression

Retry an expression until either timeout is exceeded or a condition is fullfilled.

retry(
  expr,
  envir = parent.frame(),
  until = .default_until,
  silent = TRUE,
  timeout = Inf,
  interval = 0.1
)

Arguments

expr
an expression to be evaluated, quasiquotation is supported.

envir
the environment in which the expression is to be evaluated.

until
a function of two aruments. This function is used to check if we need to re-evaluate expr. The first argument is the result of expr and the second argument is the potential error condition when expr is evaluated. The `retry` function returns the result of expr if until returns TRUE. Default value is function(res, cnd) {is.null(cnd)}. It could be also a one sided formula that is later converted to a function using rlang::as_function.

silent
suppress messages and warnings

timeout
raise an error if this amount of time in second has passed.

interval
delay between retries.

Examples

retry(10)  # returns 10 imediately

#> [1] 10

elapse <- function(s, by) {
  t <- Sys.time()
  if (t - s < by) {
    stop()
  }
  return(t - s)
}
s <- Sys.time()
retry(elapse(s, 1))  # retry until success

#> Time difference of 1.03177 secs

if (FALSE) {
  # this won't work because each retry reruns Sys.time()
  retry(elapse(Sys.time(), 1), timeout = 3)
}

# instead, we could use quasiquotation
retry(elapse(!!(Sys.time()), 1))

#> Time difference of 1.023332 secs

x <- 0
counter <- function() {
  x <<- x + 1
  x
}
retry(counter(), until = function(res, cnd) res == 10)

#> [1] 10

x <- 0
retry(counter(), until = ~ . == 10)

#> [1] 10

if (FALSE) {
  # an error is raised after 1 second
  retry(stop(), timeout = 1)

  # timeout also works for indefinite R code
  retry(while(TRUE) {}, timeout = 1)
}